Created the cutomer object in th AddCustomerController.
The observable lists for the country were causing issues.

I'm not sure if I am making a connection to the database. Tomorrow I plan to try sending the customer object to the database to see if I'm connected and to see if I can get the write (create) operation to work(thereby also testing the connection). then I will troubleshoot the read operation for the observable lists.